+++ .github/licenserc.yml
@@ -40,6 +40,5 @@ header:
     - "tidb-binlog/proto/go-binlog/secondary_binlog.pb.go"
     - "**/*.sql"
     - ".bazelversion"
-    - "tests/testmarker/testdata/"
-    - "tests/testmarker/.testmarker.yaml"
+    - "build/linter/testmarker/data.yaml"
   comment: on-failure
