+++ tests/testmarker/BUILD.bazel
@@ -3,11 +3,10 @@ load("@io_bazel_rules_go//go:def.bzl", "go_binary", "go_library", "go_test")
 go_library(
     name = "testmarker_lib",
     srcs = ["walker.go"],
-    embedsrcs = [".testmarker.yaml"],
     importpath = "github.com/pingcap/tidb/tests/testmarker",
     visibility = ["//visibility:private"],
     deps = [
-        "@com_github_ghemawat_stream//:stream",
+        "//tests/testmarker/pkg",
         "@in_gopkg_yaml_v2//:yaml_v2",
     ],
 )
@@ -25,7 +24,7 @@ go_test(
     embed = [":testmarker_lib"],
     embedsrcs = ["testdata/marker_test_go.txt"],
     deps = [
-        "@com_github_ghemawat_stream//:stream",
+        "//tests/testmarker/pkg",
         "@com_github_stretchr_testify//require",
     ],
 )
