+++ tests/testmarker/pkg/BUILD.bazel
@@ -1,0 +1,8 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "pkg",
+    srcs = ["markutil.go"],
+    importpath = "github.com/pingcap/tidb/tests/testmarker/pkg",
+    visibility = ["//visibility:public"],
+)
