--- tests/testmarker/testdata/markdata.yaml
@@ -1,2 +1,0 @@
-[]
-
